requirements.txt
playwright
pytesseract
pillow
sqlite-utils